AnswerHub/requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
asgiref==3.8.1
attrs==24.3.0
beautifulsoup4==4.12.3
cent==5.0.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.0
Django==4.2.17
django-bootstrap-v5==1.0.11
Faker==30.8.2
frozenlist==1.5.0
gunicorn==23.0.0
idna==3.10
multidict==6.1.0
packaging==24.2
pillow==11.0.0
propcache==0.2.1
psycopg2-binary==2.9.10
pycparser==2.22
pydantic==2.10.4
pydantic_core==2.27.2
PyJWT==2.10.1
python-dateutil==2.9.0.post0
requests==2.32.3
six==1.16.0
soupsieve==2.6
sqlparse==0.5.1
types-requests==2.32.0.20241016
typing_extensions==4.12.2
urllib3==2.3.0
yarl==1.18.3
